fix(content): unify selection cutover authority
